requirements.txt
requests
sentence-transformers
transformers
langchain
langchain_community
langchain-pymupdf4llm
langchain-core
langchain-milvus
langchain-huggingface
langchain-openai
pymilvus
torch
langgraph
accelerate
IPython
tqdm
mcp
asyncio
langchain_mcp_adapters
vllm
markitdown[all]
python-dotenv
numpy
sympy
mpmath